Bump reference platform to Ubuntu 18.04 LTS
Ubuntu 14.04 reached its end of life on April 30m 2019.
Let's bump the reference to the latest LTS, which is 18.04.

Ubuntu 18.04 has newer Pylint and Cppcheck, the necessary
fixes and suppresses are also included in this PR.

JerryScript-DCO-1.0-Signed-off-by: Csaba Osztrogonác oszi@inf.u-szeged.hu

+++ jerry-libm/exp.c
@@ -197,7 +197,7 @@ exp (double x) /* default IEEE double exp */
   }
   if (k >= -1021)
   {
-    ret.as_int.hi += (k << 20); /* add k to y's exponent */
+    ret.as_int.hi += (((unsigned int) k) << 20); /* add k to y's exponent */
     return ret.dbl;
   }
   else
